+++ src/test/java/racingcar/service/CarServiceTest.java
@@ -3,7 +3,6 @@ package racingcar.service;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.model.car.Car;
-import racingcar.model.car.CarDto;
 
 import java.util.Arrays;
 import java.util.List;
